% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_keyword_trend}
\alias{demo_keyword_trend}
\title{Demo}
\usage{
demo_keyword_trend(
  keyword,
  from = 1990,
  to = 2025,
  type = c("all"),
  pages = 30L,
  page_size = 20L,
  use_ai = FALSE
)
}
\arguments{
\item{keyword}{Character. Search keyword.}

\item{from}{Integer. Start year.}

\item{to}{Integer. End year.}

\item{type}{Character.}

\item{pages}{Integer. Number of pages to fetch.}

\item{page_size}{Integer. Page size (number of results per page).}

\item{use_ai}{t/f}
}
\value{
Invisibly returns the trend data frame.
}
\description{
Demo
}
