% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_law_text.R
\name{get_law_text}
\alias{get_law_text}
\title{Get full text of a BC law}
\usage{
get_law_text(law_type, law_id, collapse_whitespace = TRUE)
}
\arguments{
\item{law_type}{Character}

\item{law_id}{Character}

\item{collapse_whitespace}{Logical. If TRUE, collapse whitespace to single spaces.}
}
\value{
Character scalar containing the plain text of the law
}
\description{
Get full text of a BC law
}
