% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_trend.R
\name{keyword_trend}
\alias{keyword_trend}
\title{Keyword trend of BC laws}
\usage{
keyword_trend(keyword, from, to, type = c("all"))
}
\arguments{
\item{keyword}{Character. Search keyword.}

\item{from}{Integer. Start year}

\item{to}{Integer. End year}

\item{type}{Character.}
}
\value{
A data.frame with columns: year, law_type, count.
}
\description{
Keyword trend of BC laws
}
