% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_trend_paged.R
\name{keyword_trend_paged}
\alias{keyword_trend_paged}
\title{Keyword trend for multiple pages}
\usage{
keyword_trend_paged(
  keyword,
  from,
  to,
  type = c("all"),
  pages = 2L,
  page_size = 20L
)
}
\arguments{
\item{keyword}{Character. Search keyword.}

\item{from}{Integer. Start year}

\item{to}{Integer. End year}

\item{type}{Character}

\item{pages}{Integer. How many pages to try}

\item{page_size}{Integer.}
}
\value{
A data.frame with columns year, law_type, and count.
}
\description{
Keyword trend for multiple pages
}
