% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_search_results.R
\name{parse_search_results}
\alias{parse_search_results}
\title{Parse BC Laws search results}
\usage{
parse_search_results(xml_doc)
}
\arguments{
\item{xml_doc}{An xml_document returned by \code{\link[=search_laws]{search_laws()}}.}
}
\value{
A data.frame with one row per search hit.
}
\description{
Parse BC Laws search results
}
