% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_keyword_trend.R
\name{plot_keyword_trend}
\alias{plot_keyword_trend}
\title{Plot keyword trend}
\usage{
plot_keyword_trend(trend_data, keyword)
}
\arguments{
\item{trend_data}{A data.frame returned by keyword_trend().}

\item{keyword}{Character.}
}
\value{
Invisibly returns NULL.
}
\description{
Plot keyword trend
}
