% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_laws.R
\name{search_laws}
\alias{search_laws}
\title{Search BC laws}
\usage{
search_laws(query_string, start = 0, end = 20, n_frag = 5, l_frag = 100)
}
\arguments{
\item{query_string}{Character. Search keywords.}

\item{start}{Integer. Start index of results.}

\item{end}{Integer. End index of results.}

\item{n_frag}{Integer. Number of text fragments per hit.}

\item{l_frag}{Integer. Max length of each fragment.}
}
\value{
An xml_document with raw search results.
}
\description{
Search BC laws
}
